backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pinecone-client
unstructured
pypdf
python-docx
langchain
langchain-community
langchain-google-vertexai
langchain-pinecone
tiktoken
pandas
lxml
html5lib
beautifulsoup4
nest-asyncio
python-multipart
aiofiles
langchain-google-genai
openai
moviepy 